Move main to package
